Add files via upload
These files are a more recent version of the final step in Net Ninja REST API tutorial. All made with arrow functions and a ternary operator is used for the RED/GREEN circle that signals availability of truck in range (I'm using  load, instead of rank but the structure is the same).
If you had problems with the version on Net Ninja, these modifications  could resolves them.